Feat: ExcelService + envío de PDF y Excel por Telegram y email
- ExcelService: genera presupuesto .xlsx en memoria con Apache POI
  Cabecera autónomo, datos cliente, tabla de conceptos (concepto/tipo/
  cantidad/precio unitario/importe) y fila de totales. Estilo profesional:
  negrita, bordes, colores de cabecera azul/gris.
- UserSession: nuevos campos pendingXlsxBytes/pendingXlsxNombre
- CallbackHandler: al confirmar presupuesto genera y envía ambos ficheros
  (PDF + Excel) por Telegram; al aceptar email los adjunta los dos
- EmailService: nuevo método enviarConAdjuntos(varargs Adjunto); el método
  existente enviarConAdjunto delega en él para mantener compatibilidad
- Fix VoiceHandler: Long → int (getFileSize().intValue())
- Fix FacturaService: adaptado a PdfService.generarFactura que devuelve byte[]

+++ src/main/java/org/gestorai/bot/handlers/VoiceHandler.java
@@ -65,7 +65,7 @@ public class VoiceHandler {
         // 2. Validar metadatos del audio ANTES de descargarlo
         Voice voice = message.getVoice();
         int duracion = voice.getDuration()  != null ? voice.getDuration()  : 0;
-        int tamano   = voice.getFileSize()  != null ? voice.getFileSize()  : 0;
+        int tamano   = voice.getFileSize()  != null ? voice.getFileSize().intValue()  : 0;
 
         RateLimiter.Resultado resultado = rateLimiter.comprobarAudio(telegramId, duracion, tamano);
         if (resultado != RateLimiter.Resultado.OK) {
